Fixed Issue 265 : Closing and reopening of parent project in workspace causes error http://code.google.com/p/q4e/issues/detail?id=265
- In EclipseMavenArtifactResolver, if an artifact matches a workspace project, 
the file of that artifact is set to the pom file of the maven project. This can 
cause some code to break if they assume that the file that will be returned by 
getFile() of an artifact is its jar file. The fix is to implement the 
getPath(IMavenArtifact) of LocalRepository so it can be used to query the location 
of the jar and use this instead of the artifact's getFile() in constructing 
jar dependencies.