Completed  Issue 271: MavenProject.getArtifacts() might contain stale information http://code.google.com/p/q4e/issues/detail?id=271
- Modified UpdateClasspathJob so we could attach an IJobChangeListener 
to the classpath update execution. This will allow us to add code that 
will clear the pom error markers that the exception handler has created before 
the job executes.
- Uses the artifacts of a maven project instead of its calculated classpath 
entries to check whether it has a dependency on the opened/closed project.
- Set the project as modified in the maven project cache if it has a dependency 
on the closed/opened project. This will solve Issue 271 because the next time an 
IMavenProject instance of this project is needed, it will be taken from the 
embedder and our artifact resolver will be executed again.